refactor: delete unnecessary folders and csss file for language selector
